requirements.txt
discord.py
python-dotenv
discord-ext-commands
asyncio
openai
questionary
